ADD part of the functions to db_interface com main

+++ main.py
@@ -7,7 +7,7 @@ from telebot.types import ReplyKeyboardMarkup, KeyboardButton, Message
 import constants as const
 import random
 import processing as pr
-import db_interface_test
+import db_interface
 
 # Initialize the bot using the bot token
 bot = telebot.TeleBot(f"{const.API_KEY_HOSTED}")
@@ -32,7 +32,7 @@ class Poll:
 def start_handler(message):
     menu_keyboard = ReplyKeyboardMarkup(row_width=1)
     menu_keyboard.add(KeyboardButton('/help'))
-    reply_text = db_interface_test.store(message.chat.id, "user", 0) #change it -сделано---------------------------------------------
+    reply_text = db_interface.userRegistration(message.chat.id) #change it -сделано---------------------------------------------
     bot.reply_to(message, reply_text, reply_markup=menu_keyboard)
 
 
@@ -41,7 +41,6 @@ def start_handler(message):
 def help_handler(message):
     bot.reply_to(message,
                  'Type:\n"/quiz" - to get a quiz\n'
-                 '"/add_word" - to add a new word\n'
                  '"/whole_dict" - to check all the words\n'
                  '"/start_mailing" - to start getting quizzes\n'
                  '"/stop_mailing" - to stop mailing\n'
@@ -49,73 +48,60 @@ def help_handler(message):
                  '"/game" - to get a game\n')
 
 
-# Define a function to handle the /whole_dict command
 @bot.message_handler(commands=['whole_dict'])
 def whole_dict_handler(message):
-    t = time.time()
-    bot.send_message(chat_id=message.chat.id, text="Генерирую файл...")
-    dictionary = db_interface_test.get_words()[1] #change it ---------------------------------------------
-    file_path = "./cache/output.txt"
+    # t = time.time()
+    # bot.send_message(chat_id=message.chat.id, text="Генерирую файл...")
+    # dictionary = db_interface.get_all_words() #change it ---------------------------------------------
+    # file_path = "./cache/output.txt"
 
-    with open(file_path, "w", encoding="utf-8") as file:
-        for word in dictionary:
-            file.write(f"{word['word']}-{word['translation']};\n")
+    # with open(file_path, "w", encoding="utf-8") as file:
+    #     for word in dictionary:
+    #         file.write(f"{word['word']}-{word['translation']};\n")
 
-    bot.send_document(chat_id=message.chat.id, document=open(file_path, "rb"))
-    os.remove(file_path)
-    bot.send_message(chat_id=message.chat.id, text="Файл успешно сгенерирован")
+    # bot.send_document(chat_id=message.chat.id, document=open(file_path, "rb"))
+    # os.remove(file_path)
+    # bot.send_message(chat_id=message.chat.id, text="Файл успешно сгенерирован")
 
-    print(time.time() - t, "out")
-
-# this functionality has to be removed for some time
-# adds word to the dictionary.json file
-@bot.message_handler(commands=['add_word'])
-def add_word(message):
-    bot.reply_to(message, 'Введите новое слово и перевод в формате "слово1-перевод1;слово2-перевод2;и тд;'
-                          '\nтакже, пожалуйста, вводите слова в единственном числе и мужском роде"')
-    bot.register_next_step_handler(message, add_and_verify)
-
-
-# Добавляет слова непосредственно в словарь
-def add_and_verify(message):
-    bot.send_message(message.chat.id, "Добавляю...")
-    listOfNewWords = pr.prepare_text(message.text)
-    db_interface_test.add_word_to_bd(listOfNewWords, message.chat.id) # change it -------------------------
-    bot.send_message(message.chat.id, "Словарь обновлен!")
+    # print(time.time() - t, "out")
+    bot.send_message(chat_id=message.chat.id, text="Временно недоступно")
 
 
 # Add words from files to the dict
 @bot.message_handler(content_types=['document'])
-def add_word_from_file(message):
-    file_info = message.document
-    file_id = file_info.file_id
-    file_name = file_info.file_name
+def add_dictionary_from_file(message):
+    if db_interface.get_user_access(message.chat.id) == 'user':
+        bot.reply_to(message, "Ваш уровень доступа не позволяет добавлять новый словарь.")
+    else: 
+        file_info = message.document
+        file_id = file_info.file_id
+        file_name = file_info.file_name
 
-    # Запрос файла с использованием его file_id
-    file_info = bot.get_file(file_id)
-    downloaded_file = bot.download_file(file_info.file_path)
+        # Запрос файла с использованием его file_id
+        file_info = bot.get_file(file_id)
+        downloaded_file = bot.download_file(file_info.file_path)
 
-    # Сохранение файла локально
-    with open(file_name, 'wb') as new_file:
-        new_file.write(downloaded_file)
+        # Сохранение файла локально
+        with open(file_name, 'wb') as new_file:
+            new_file.write(downloaded_file)
 
-    # Чтение содержимого файла
-    with open(file_name, 'r', encoding='utf-8') as file:
-        file_content = file.read()
+        # Чтение содержимого файла
+        with open(file_name, 'r', encoding='utf-8') as file:
+            file_content = file.read()
 
-    os.remove(file_name)
+        os.remove(file_name)
 
-    bot.send_message(message.chat.id, "Добавляю...")
+        bot.send_message(message.chat.id, "Добавляю...")
 
-    listOfNewWords = pr.prepare_text(file_content)
-    db_interface_test.add_word_to_bd(listOfNewWords, message.chat.id) # change it ----------------------------
+        new_dictionary = pr.prepare_text(file_content)
+        text = db_interface.add_new_dictionary(new_dictionary, 'TEST')
 
-    bot.reply_to(message, "Словарь обновлен")
+        bot.reply_to(message, text)
 
 
 # generates quiz when user types "/quiz"
 def generate_quiz():
-    dictionary = db_interface_test.get_words()[0][random.randint(0, 4)] # change it --------------------------------------------------------
+    dictionary = db_interface.get_words()[0][random.randint(0, 4)] # change it --------------------------------------------------------
     answer_options = random.sample(dictionary, 4)
     word_number = random.randint(0, 3)
     print(answer_options)  # for debugging
@@ -140,24 +126,26 @@ def create_poll():
 # sends quiz
 @bot.message_handler(commands=['quiz'])
 def send_quiz(MesOrNum, need_list=None):
-    if need_list is None:
-        need_list = []
+    # if need_list is None:
+    #     need_list = []
 
-    if isinstance(MesOrNum, int):
-        chat_id = MesOrNum
-    else:
-        chat_id = MesOrNum.chat.id
-        need_list.append(chat_id)
+    # if isinstance(MesOrNum, int):
+    #     chat_id = MesOrNum
+    # else:
+    #     chat_id = MesOrNum.chat.id
+    #     need_list.append(chat_id)
 
-    poll = create_poll()
-    for chat_id1 in need_list:
-        poll.send(chat_id1, bot)
+    # poll = create_poll()
+    # for chat_id1 in need_list:
+    #     poll.send(chat_id1, bot)
+    bot.send_message(chat_id=MesOrNum.chat.id, text="Временно недоступно")
 
 
 # function to send quizzes to the users
 def check_send_quiz():
-    need_list = db_interface_test.get_needed_users()
-    send_quiz(0, need_list)
+    # need_list = db_interface.get_needed_users()
+    # send_quiz(0, need_list)
+    pass
 
 
 # checks if there are users asking for quiz
@@ -185,8 +173,9 @@ def get_valid_integer(text):
 # updates user's mailing status
 @bot.message_handler(commands=['start_mailing'])
 def start_mailing(message):
-    bot.send_message(message.chat.id, "Введите, как часто Вы хотите, чтобы приходили квизы(в минутах)")
-    bot.register_next_step_handler(message, start_mailing_time)
+    # bot.send_message(message.chat.id, "Введите, как часто Вы хотите, чтобы приходили квизы(в минутах)")
+    # bot.register_next_step_handler(message, start_mailing_time)
+    bot.send_message(chat_id=message.chat.id, text="Временно недопустимая команда")
 
 
 def start_mailing_time(message):
@@ -202,10 +191,10 @@ def start_mailing_time(message):
         return
 
     # запуск рассылки, время переводится в секунды
-    f = db_interface_test.started_mailing(message.chat.id)
+    f = db_interface.started_mailing(message.chat.id)
     if f == 0:
         bot.send_message(message.chat.id, "Запустил рассылку")
-        db_interface_test.update_mailing(message.chat.id, minutes)
+        db_interface.update_mailing(message.chat.id, minutes)
 
     else:
         bot.send_message(message.chat.id, "У Вас уже запущена рассылка")
@@ -214,20 +203,22 @@ def start_mailing_time(message):
 # Stops mailing
 @bot.message_handler(commands=['stop_mailing'])
 def stop_mailing(message):
-    f = db_interface_test.started_mailing(message.chat.id)
-    if f != 0:
-        # t.cancel()
-        bot.send_message(message.chat.id, "Остановил рассылку")
-        db_interface_test.update_mailing(message.chat.id, 0)
-    else:
-        bot.send_message(message.chat.id, "У Вас не запущена рассылка")
+    # f = db_interface.started_mailing(message.chat.id)
+    # if f != 0:
+    #     # t.cancel()
+    #     bot.send_message(message.chat.id, "Остановил рассылку")
+    #     db_interface.update_mailing(message.chat.id, 0)
+    # else:
+    #     bot.send_message(message.chat.id, "У Вас не запущена рассылка")
+    bot.send_message(chat_id=message.chat.id, text="Временно недопустимая команда")
 
 
 # Changes a period of mailing
 @bot.message_handler(commands=['change_mailing_time'])
 def change_mailing_time_0(message):
-    bot.reply_to(message, 'Введите, как часто Вы хотите, чтобы приходили квизы(в минутах)')
-    bot.register_next_step_handler(message, change_mailing_time)
+    # bot.reply_to(message, 'Введите, как часто Вы хотите, чтобы приходили квизы(в минутах)')
+    # bot.register_next_step_handler(message, change_mailing_time)
+    bot.send_message(chat_id=message.chat.id, text="Временно недопустимая команда")
 
 
 def change_mailing_time(message):
@@ -242,12 +233,12 @@ def change_mailing_time(message):
         bot.register_next_step_handler(message, change_mailing_time)
         return
 
-    f = db_interface_test.started_mailing(message.chat.id)
+    f = db_interface.started_mailing(message.chat.id)
 
     if f != 0:
         # t.cancel()
         bot.send_message(message.chat.id, "Изменил время")
-        db_interface_test.update_mailing(message.chat.id, minutes)
+        db_interface.update_mailing(message.chat.id, minutes)
     else:
         bot.send_message(message.chat.id, "У Вас не запущена рассылка")
 
@@ -255,8 +246,9 @@ def change_mailing_time(message):
 # Creates a game for users
 @bot.message_handler(commands=['game'])
 def game_get_num(message):
-    bot.reply_to(message, 'Введите количество квизов, которое вы хотите получить')
-    bot.register_next_step_handler(message, game)
+    # bot.reply_to(message, 'Введите количество квизов, которое вы хотите получить')
+    # bot.register_next_step_handler(message, game)
+    bot.send_message(chat_id=message.chat.id, text="Временно недопустимая команда")
 
 
 def game(message):
@@ -303,7 +295,7 @@ def game(message):
 
 
 print(__name__)
-set_interval(check_send_quiz, 60)
+# set_interval(check_send_quiz, 60)
 
 if __name__ == '__main__':
     bot.polling()
